Add a RESET button to the leaflet widget.
Sometimes the  widget map gets all screwed up for whatever reason.  Adding a "RESET"
button helps the user by giving her an easy way to return to the default
settings.

This issue is related to nucivic/usda-nal#765.  This commits adds a possible way
of dealing with the issue but acknowledges that there is no for this module to implement a general
automatic fix for this problem because the module has no way of knowing how the
leaflet widget will be wrapped under all circumstances.