Improved multiple inputs tabs implementation for leaflet widget (#10)
* Improve tabs implementation for leaflet widget

* Update the field's input when updating the map.

* Fix missing geojson propreties.

* Fix too generic jQuery.once class name

* Use https for the map tile if available

backport from "nucivic_northd_new_version" branch.

* Pull draw.js from USDA master

Seems to be working well in production.

* Add reset button to Map.

Sometimes the  widget map gets all screwed up for whatever reason.  Adding a "RESET"
button helps the user by giving her an easy way to return to the default
settings.

This issue is related to NuCivic/usda-nal#765.  This commits adds a possible way
of dealing with the issue but acknowledges that there is no for this module to implement a general
automatic fix for this problem because the module has no way of knowing how the
leaflet widget will be wrapped under all circumstances.

* Fix cannot delete point among others.

There are multiple issues fixed with this commit.

- [ ] The user could not delete a point after creating it [Ref
  nucivic/usda-nal#751]

- [ ] The user could not add more than one point or polygon (feature).
- [ ] The user could not edit the position of a feature.
- [ ] None of the binding functions for the tabs were working.
- [ ] The reset functionality can reset even after an edit has occurred.

* Clean up syntax.

* Removing reset button for now.

Reset is better implemented as an extension of leaflet.draw.

Ref #751.

* Remove reset and make adding points easier.

Reset should be implemented via extension of leaflet.draw.

Does not hide the map when adding points so that it's easier to follow what is going on.

* Allow multiple point entries at once.

Ref nucivic/usda-nal#752

This commit let's the user enter multiple points delimited using semicolons.

* Use correct latitude/longitude order.

Ref nucivic/usda-nal#752.

* Update draw.js

* Drop debug statement

* Set L.Icon.Default.imagePath if missing

* Pull from NuCivic/leaflet_draw_widget/pull/17

* Append missing "/" to image path